refactor!: remove libp2p.keychain (#2084)
Extract the keychain into it's own package.
