+++ packages/libp2p/package.json
@@ -124,7 +124,6 @@
     "@libp2p/crypto": "^2.0.7",
     "@libp2p/interface": "^0.1.5",
     "@libp2p/interface-internal": "^0.1.8",
-    "@libp2p/keychain": "^3.0.7",
     "@libp2p/logger": "^3.0.5",
     "@libp2p/multistream-select": "^4.0.5",
     "@libp2p/peer-collections": "^4.0.7",
